requirements.txt
django==5.0.7
djangorestframework==3.15.2
polars==1.4.0
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
openpyxl==3.1.5